Added Permission control

+++ routes/api/bouquetRoute.js
@@ -3,8 +3,12 @@ var bManager=require('../../db/bouquetsManager')();
 var loginManager=require('../../db/loginManager')();
 let userType=require('../../lib/constants').UserType;
 
+let PermissionManager = require('../../lib/PermissionManager');
+
 router.all(/^\/(.*)/, (req,resp,next)=>{
     console.log(req.user);
+    console.log(req.roles);
+
     if(req.user){
         next();
     }else{
@@ -13,6 +17,12 @@ router.all(/^\/(.*)/, (req,resp,next)=>{
 })
 
 router.get('/',(req,res)=>{
+
+    if(!PermissionManager.canReadBouquets(req.roles.grantLevel)){
+        resp.status(400).json("Not enough permission");
+        return;
+    }
+
     bManager.readAll(req.dbSession)
     .then((result) => {
         res.json({
@@ -31,8 +41,11 @@ router.get('/',(req,res)=>{
 })
 
 router.put('/',(req,resp)=>{
-
     if(req.isSystem){ 
+        if(! PermissionManager.canCreateBouquet(req.roles.grantLevel)){
+            resp.status(400).json("Not enough permission");
+            return;
+        }
         /// TODO: check if the user has sufficient rights to create a bouquet
         bManager.create(req.dbSession,req.body)
         .then((result) => {
@@ -54,8 +67,12 @@ router.put('/',(req,resp)=>{
 })
 
 router.patch('/:id',(req,resp)=>{
-    console.log(req.params.id);
     if(req.params.id && req.isSystem){
+        if(! PermissionManager.canCreateBouquet(req.roles.grantLevel)){
+            resp.status(400).json("Not enough permission");
+            return;
+        }
+
         bManager.patch(req.dbSession,req.params.id,req.body)
         .then(()=>{
             resp.json({success:true});
@@ -71,6 +88,11 @@ router.options('/:id',(req,resp)=>{
     console.log(req.body);
 
     if(req.params.id && req.isSystem && req.body.option){
+        if(! PermissionManager.canLockBouquet(req.roles.grantLevel)){
+            resp.status(400).json("Not enough permission");
+            return;
+        }
+
         let opt=req.body.option;
         if(opt==='lock'){
             bManager.lock(req.dbSession,req.params.id)
@@ -98,6 +120,11 @@ router.options('/:id',(req,resp)=>{
 
 router.delete('/:id',(req,resp)=>{
     if(req.params.id && req.isSystem){
+        if(! PermissionManager.canDeleteBouquet(req.roles.grantLevel)){
+            resp.status(400).json("Not enough permission");
+            return;
+        }
+
         bManager.delete(req.dbSession,req.params.id)
         .then(()=>{
             resp.json({success:true});
